% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiDMFMC.R
\name{SemiDMFMC}
\alias{SemiDMFMC}
\title{[S]emi-parametirc [D]ynamic [M]ulti-[F]actor [M]ulti-[C]umulant estimation}
\usage{
SemiDMFMC(
  X,
  ff,
  Z = NULL,
  SemiFL = F,
  sel.bw = "uni",
  con_residual = T,
  best_model = NULL,
  Penalty = c("NONE", "LASSO", "MCP", "SCAD"),
  factor.control = list(var.model = "sGARCH", var.targeting = F, var.distribution =
    "sged", Corr.Struture = c("ica", "dcc", "copula"), dcc.model = "DCC", copula.model =
    list(copula = "mvt", method = "ML", time.varying = FALSE, transformation = "spd"),
    tgc.type = "leverage", tgc.targeting = F, mean.model = list(armaOrder = c(1, 1)),
    CTGC = FALSE, rep_sim_f = 10),
  eps.control = list(var.model = "sGARCH", var.targeting = F, var.distribution =
    "sged", tgc.type = "leverage", tgc.targeting = F, mean.model = list(armaOrder = c(0,
    0)), rep_sim_e = 10),
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{ff}{The factors drive the data X.}

\item{Z}{The variable drive factor loadings.}

\item{SemiFL}{Logical. If \code{TRUE}, the factor loadings will be semi-parametric function.}

\item{sel.bw}{Bandwith selection method.}

\item{con_residual}{Logical, whether the moments of the residuals are time-varying or not.}

\item{best_model}{A vector. It gives which factor loadings are semi-parametric.}

\item{Penalty}{Only used for multi-factor model. Penalty regression to estimate factor loadings.}

\item{factor.control}{The parameters of factors' moments estimation.}

\item{eps.control}{The parameters of errors' moments estimation.}

\item{...}{Any other passthru parameters.}
}
\value{
Estimated covariance, co-skewness, co-exkurtosis, co-kurtosis and the results of regression.
}
\description{
[S]emi-parametirc [D]ynamic [M]ulti-[F]actor [M]ulti-[C]umulant estimation
}
