% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiDMFMC.R
\name{SemiDMFMC}
\alias{SemiDMFMC}
\title{[S]emi-parametirc [D]ynamic [M]ulti-[F]actor [M]ulti-[C]umulant estimation}
\usage{
SemiDMFMC(
  X,
  ff,
  Z = NULL,
  SemiFL = F,
  sel.bw = "uni",
  con_residual = T,
  best_model = NULL,
  Penalty = c("NONE", "LASSO", "MCP", "SCAD"),
  tgc.type = c("linear", "leverage", "n-leverage"),
  rep_sim = 10,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame with t rows (samples) and n columns (variables).}

\item{ff}{The factors drive the data X.}

\item{Z}{The variable drive factor loadings.}

\item{SemiFL}{Logical. If \code{TRUE}, the factor loadings will be semi-parametric function.}

\item{sel.bw}{Bandwith selection method.}

\item{con_residual}{Logical, whether the moments of the residuals are time-varying or not.}

\item{best_model}{A vector. It gives which factor loadings are semi-parametric.}

\item{Penalty}{Only used for multi-factor model. Penalty regression to estimate factor loadings.}

\item{tgc.type}{the time varying type of TGC distribution.}

\item{rep_sim}{Repeat times for estimating time-varying parameters.}

\item{...}{Any other passthru parameters}
}
\value{
Estimated covariance, co-skewness, co-exkurtosis, co-kurtosis and the results of regression.
}
\description{
[S]emi-parametirc [D]ynamic [M]ulti-[F]actor [M]ulti-[C]umulant estimation
}
